refactor: refactor package structure from layer-based to domain-based
