+++ src/main/java/com/githubsalt/omoib/user/UserRepository.java
@@ -1,6 +1,5 @@
-package com.githubsalt.omoib.repository;
+package com.githubsalt.omoib.user;
 
-import com.githubsalt.omoib.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
